Divide Upload container files into web and native. Move queries and mutations out of the Upload container
